+++ WalletWasabi.Fluent/ViewModels/Wallets/Buy/Workflows/IChatMessageEditor.cs
@@ -1,0 +1,10 @@
+using WalletWasabi.BuyAnything;
+
+namespace WalletWasabi.Fluent.ViewModels.Wallets.Buy.Workflows;
+
+public interface IChatMessageEditor
+{
+	bool IsEditable(ChatMessage chatMessage);
+
+	IWorkflowStep? GetEditor(ChatMessage chatMessage);
+}
